Create README.md
The project I determined to proceed with is, based on analyzing the data collected by the Internet of Medical Things(IoMT) devices. I am solely focusing on the problem statement 'Can human heart rate and  SPo2 levels predict the potential occurrence of diseases in the future? The data is collected using the human fitness trackers connected via IoT(Internet of things).To solve this problem statement I used a different set of models linked with various predictor variables.